% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codified-package.R
\docType{package}
\name{codified-package}
\alias{codified}
\alias{codified-package}
\title{codified: Produce standard/formalized demographics tables}
\description{
Augment clinical data with metadata to create
    output used in conventional publications and reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OuhscBbmc/codified}
  \item \url{https://github.com/higgi13425/nih_enrollment_table}
  \item Report bugs at \url{https://github.com/OuhscBbmc/codified/issues}
}

}
\author{
\strong{Maintainer}: Will Beasley \email{wibeasley@hotmail.com} (0000-0002-5613-5006)

Other contributors:
\itemize{
  \item Peter Higgins [contributor]
}

}
\keyword{internal}
