% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-nih-enrollment.R
\name{table_nih_enrollment}
\alias{table_nih_enrollment}
\alias{table_nih_enrollment_pretty}
\title{Produce an NIH-compliant enrolment table.}
\usage{
table_nih_enrollment(d, d_lu_gender = NULL, d_lu_race = NULL,
  d_lu_ethnicity = NULL)
}
\arguments{
\item{d}{\link{data.frame} of observed values in the investigation. Required.}

\item{d_lu_gender}{\link{data.frame} that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.}

\item{d_lu_race}{\link{data.frame} that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.}

\item{d_lu_ethnicity}{\link{data.frame} that maps the observed levels of gender to the NIH-recommended levels of gender. Required only if the levels are not the same.}
}
\value{
Table for publication
}
\description{
Produce an NIH enrollment table, leveraging metadata to adapt to the observed \link{data.frame}.
}
\details{
https://grants.nih.gov/grants/how-to-apply-application-guide/forms-d/general/g.500-phs-inclusion-enrollment-report.htm
}
\examples{
ds_1 <- tibble::tribble(
  ~subject_id,   ~gender  , ~race                      ,   ~ethnicity                     ,
           1L,   "Male"   , "Black or African American",  "Not Hispanic or Latino"        ,
           2L,   "Male"   , "Black or African American",  "Not Hispanic or Latino"        ,
           3L,   "Female" , "Black or African American",  "Unknown/Not Reported Ethnicity",
           4L,   "Male"   , "White"                    ,  "Not Hispanic or Latino"        ,
           5L,   "Male"   , "White"                    ,  "Not Hispanic or Latino"        ,
           6L,   "Female" , "White"                    ,  "Not Hispanic or Latino"        ,
           7L,   "Male"   , "White"                    ,  "Hispanic or Latino"            ,
           8L,   "Male"   , "White"                    ,  "Hispanic or Latino"
)

table_nih_enrollment(ds_1)
table_nih_enrollment_pretty(ds_1)

\dontrun{
table_nih_enrollment(ds_1) \%>\%
  tidyr::spread(key=gender, value=n)

table_nih_enrollment(ds_1) \%>\%
  dplyr::mutate(
    gender_ethnicity = paste0(gender, " by ", ethnicity)
  ) \%>\%
  dplyr::select(-gender, -ethnicity) \%>\%
  tidyr::spread(key=gender_ethnicity, value=n)
  }

ds_2 <- tibble::tribble(
  ~subject_id,  ~gender , ~race                      , ~ethnicity    ,
           1L,  "Male"  , "Black or African American", "Not Latino"  ,
           2L,  "Male"  , "Black or African American", "Not Latino"  ,
           3L,  "Female", "Black or African American", "Unknown"     ,
           4L,  "Male"  , "White"                    , "Not Latino"  ,
           5L,  "Male"  , "White"                    , "Not Latino"  ,
           6L,  "Female", "White"                    , "Not Latino"  ,
           7L,  "Male"  , "White"                    , "Latino"      ,
           8L,  "Male"  , "White"                    , "Latino"
)

ds_lu_ethnicity <- tibble::tribble(
  ~input      ,   ~displayed                      ,
  "Not Latino",  "Not Hispanic or Latino"         ,
  "Latino"    ,  "Hispanic or Latino"             ,
  "Unknown"   ,  "Unknown/Not Reported Ethnicity"
)
table_nih_enrollment(ds_2, d_lu_ethnicity=ds_lu_ethnicity)
table_nih_enrollment_pretty(ds_2)

}
\author{
Will Beasley, Peter Higgins, Andrew Peters, Sreeharsha Mandem
}
